牛逼

+++ MMS/src/main/java/com/example/demo/mapper/receptorMapper.java
@@ -9,7 +9,7 @@ public interface receptorMapper {
 
     int insert(receptor record);
 
-    receptor selectByPrimaryKey(Long id);
+    receptor selectByPrimaryKey(Integer id);
 
     receptor selectByMedicareId(Long medicareId);
